move files in entropy plugin to main aspect
