+++ include/aspect/adiabatic_conditions/compute_entropy_profile.h
