+++ include/aspect/material_model/entropy_model.h
