+++ source/material_model/rheology/visco_plastic.cc
@@ -809,6 +809,8 @@ namespace aspect
   namespace MaterialModel
   {
 #define INSTANTIATE(dim) \
+  template class PlasticAdditionalOutputs<dim>; \
+  \
   namespace Rheology \
   { \
     template class ViscoPlastic<dim>; \
