+++ tests/entropy_conduction.cc
@@ -18,7 +18,4 @@
   <http://www.gnu.org/licenses/>.
 */
 
-#include "../benchmarks/entropy_adiabat/plugins/compute_entropy_profile.cc"
 #include "../benchmarks/entropy_adiabat/plugins/entropy_advection.cc"
-#include "../benchmarks/entropy_adiabat/plugins/entropy_model.cc"
-#include "../benchmarks/entropy_adiabat/plugins/prescribed_temperature.cc"
